// Features/reports/Types/IConsolidatedDataForReport.ts
export interface IConsolidatedDataForReport {
    identifier: string;
    idCostCenter: number;
    costCenterName: string;
    idEmployee: number;
    employeeName: string;
    extraHourDate: string;
    extraHourAmount: number;
    idPayrollPay: number;
    payrollName: string;
    idConcept: number;
    concept: string;
    idAccountingAccount: number;
    accountNumber: string;
    name: string;
    conceptCode: string;
    idCompany: number;
}